--- .classpath
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry exported="true" kind="lib" path="lib/poi-ooxml-schemas-3.11.jar"/>
-	<classpathentry exported="true" kind="lib" path="lib/xmlbeans-2.6.0.jar"/>
-	<classpathentry exported="true" kind="lib" path="lib/poi-3.11.jar"/>
-	<classpathentry exported="true" kind="lib" path="lib/poi-ooxml-3.11.jar"/>
-	<classpathentry exported="true" kind="lib" path="lib/fastjson-1.1.46.jar"/>
-	<classpathentry kind="con" path="org.eclipse.pde.core.requiredPlugins"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8"/>
-	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="output" path="bin"/>
-</classpath>
