--- build.properties
@@ -1,10 +1,0 @@
-source.. = src/
-output.. = bin/
-bin.includes = plugin.xml,\
-               META-INF/,\
-               .,\
-               lib/fastjson-1.1.46.jar,\
-               lib/poi-ooxml-3.11.jar,\
-               lib/poi-3.11.jar,\
-               lib/xmlbeans-2.6.0.jar,\
-               lib/poi-ooxml-schemas-3.11.jar
