+++ code.parser.excel/META-INF/MANIFEST.MF
@@ -1,0 +1,19 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Excel
+Bundle-SymbolicName: code.parser.excel;singleton:=true
+Bundle-Version: 1.0.3.qualifier
+Bundle-Activator: code.parser.excel.Activator
+Require-Bundle: org.eclipse.ui,
+ org.eclipse.core.runtime,
+ code.parser;bundle-version="1.0.2",
+ org.apache.commons.io,
+ org.apache.commons.lang,
+ org.eclipse.core.resources
+Bundle-RequiredExecutionEnvironment: JavaSE-1.8
+Bundle-ActivationPolicy: lazy
+Bundle-ClassPath: lib/poi-3.11.jar,
+ lib/poi-ooxml-3.11.jar,
+ lib/poi-ooxml-schemas-3.11.jar,
+ lib/xmlbeans-2.6.0.jar,
+ .
