+++ code.parser.excel/build.properties
@@ -1,0 +1,9 @@
+source.. = src/
+output.. = bin/
+bin.includes = META-INF/,\
+               .,\
+               plugin.xml,\
+               lib/poi-3.11.jar,\
+               lib/poi-ooxml-3.11.jar,\
+               lib/poi-ooxml-schemas-3.11.jar,\
+               lib/xmlbeans-2.6.0.jar
