+++ code.parser.excel/plugin.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?eclipse version="3.4"?>
+<plugin>
+   <extension
+         point="org.eclipse.ui.popupMenus">
+      <objectContribution
+            adaptable="true"
+            id="code.parser.excel.contribution1"
+            nameFilter="*.java"
+            objectClass="org.eclipse.core.resources.IFile">
+             <action
+               class="code.parser.excel.popup.actions.GenerateJavaExcel"
+               id="code.parser.excel.newAction"
+               label="Generate Java Excel">
+         </action>
+      </objectContribution>
+   </extension>
+
+</plugin>
