+++ code.parser/META-INF/MANIFEST.MF
@@ -1,0 +1,21 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Parser
+Bundle-SymbolicName: code.parser;singleton:=true
+Bundle-Version: 1.0.3.qualifier
+Bundle-Activator: code.parser.Activator
+Require-Bundle: org.eclipse.ui,
+ org.eclipse.core.runtime,
+ org.eclipse.core.resources,
+ org.eclipse.jdt.core,
+ org.apache.commons.io;bundle-version="2.0.1",
+ org.apache.commons.lang;bundle-version="2.6.0",
+ org.apache.velocity;bundle-version="1.5.0"
+Bundle-RequiredExecutionEnvironment: JavaSE-1.8
+Bundle-ActivationPolicy: lazy
+Bundle-ClassPath: lib/fastjson-1.1.46.jar,
+ .
+Export-Package: code.parser.convertor,
+ code.parser.dialog,
+ code.parser.util,
+ com.alibaba.fastjson
