+++ code.parser/plugin.xml
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?eclipse version="3.4"?>
+<plugin>
+
+   <extension
+         point="org.eclipse.ui.popupMenus">
+      <objectContribution
+            adaptable="true"
+            id="code.parser.contribution1"
+            nameFilter="*.java"
+            objectClass="org.eclipse.core.resources.IFile">
+         <action
+               class="code.parser.popup.actions.GenerateJavaJson"
+               id="code.parser.newAction"
+               label="Generate Java Json">
+         </action>
+<!--         <action
+               class="code.parser.popup.actions.GenerateJavaXls"
+               id="code.parser.newAction"
+               label="Generate Java xls">
+         </action>
+-->      </objectContribution>
+      <objectContribution
+            adaptable="true"
+            id="code.parser.contribution1"
+            nameFilter="*.json"
+            objectClass="org.eclipse.core.resources.IFile">
+         <action
+               class="code.parser.popup.actions.GenerateJavaCode"
+               id="code.parser.newAction"
+               label="Generate Java Code">
+         </action>
+      </objectContribution>
+   </extension>
+
+</plugin>
