+++ code.parser/src/config.properties
@@ -1,0 +1,2 @@
+#javadoc.tags=@inMsg,@outMsg
+javadoc.tags=@author,@param,@return,@throws
