--- src/config.properties
@@ -1,2 +1,0 @@
-#javadoc.tags=@inMsg,@outMsg
-javadoc.tags=@author,@param,@return,@throws
